Refactor pending transactions
