+++ apps/block_scout_web/lib/block_scout_web/controllers/pending_transaction_controller.ex
@@ -6,6 +6,7 @@ defmodule BlockScoutWeb.PendingTransactionController do
 
   alias BlockScoutWeb.{Controller, TransactionView}
   alias Explorer.Chain
+  alias Explorer.Chain.Transaction
   alias Phoenix.View
 
   {:ok, burn_address_hash} = Chain.string_to_address_hash(burn_address_hash_string())
@@ -64,7 +65,7 @@ defmodule BlockScoutWeb.PendingTransactionController do
   end
 
   defp get_pending_transactions_and_next_page(options) do
-    transactions_plus_one = Chain.recent_pending_transactions(options, true)
+    transactions_plus_one = Transaction.recent_pending_transactions(options, true)
     split_list_by_page(transactions_plus_one)
   end
 end
